Split hemi-earn utils into per-concern modules
The _utils barrel had grown to 34 exports in a single file, with a
1593-line test file alongside it, making it hard to tell where a helper
belonged or which tests covered it. decodeFailureReason and earnedAmount
already lived in their own sibling modules, so this extends that layout
to the rest.

Helpers are grouped by concern and layered so imports only ever point
downwards: hashes and pools have no local dependencies, settlement
builds on hashes, transactionPredicates on settlement, and settleUi and
earnRows sit on top. That ordering is what keeps a predicates and
settlement import cycle from forming.

Every function moved byte for byte, and each module got its own test
file carrying the original describe blocks unchanged. Four redundant
return type annotations were dropped where TypeScript already infers
the same type, and the 785 tests pass unchanged.
