+++ portal/app/[locale]/hemi-earn/_components/transactionsSection/withdrawStatusCell.tsx
@@ -12,7 +12,7 @@ import { useAccount } from 'wagmi'
 import { useCooldownDuration } from '../../_hooks/useCooldownDuration'
 import { useEarnPools } from '../../_hooks/useEarnPools'
 import { useIsCooldownEligible } from '../../_hooks/useIsCooldownEligible'
-import { findPoolByAsset } from '../../_utils'
+import { findPoolByAsset } from '../../_utils/pools'
 import { useEarnCooldownRemaining } from '../../pool/[shareAddress]/_hooks/useEarnCooldownRemaining'
 import { type EarnPool, type EarnTransaction } from '../../types'
