+++ portal/app/[locale]/hemi-earn/_utils/formatApy.ts
@@ -1,0 +1,8 @@
+import { formatPercentage } from 'utils/format'
+
+export const formatApyDisplay = function (apy: number) {
+  if (apy !== 0 && apy > -0.01 && apy < 0.01) {
+    return apy > 0 ? '< 0.01%' : '< -0.01%'
+  }
+  return formatPercentage(apy)
+}
