+++ portal/app/[locale]/hemi-earn/_utils/hashes.ts
@@ -1,0 +1,2 @@
+export const hashesMatch = (a: string | undefined, b: string | undefined) =>
+  !!a && !!b && a.toLowerCase() === b.toLowerCase()
