--- portal/test/app/[locale]/hemi-earn/_utils/index.test.ts
@@ -1,1593 +1,0 @@
-import { ProgressStatus } from 'components/reviewOperation/progressStatus'
-import { mainnet } from 'networks/mainnet'
-import { type EvmToken } from 'types/token'
-import { type Address, type Hash, zeroAddress } from 'viem'
-import { describe, expect, it } from 'vitest'
-
-import {
-  canRetryRow,
-  claimRecoverSettlement,
-  enrichWithSettlement,
-  findLocalSettlement,
-  findPoolByAsset,
-  findPoolByShare,
-  formatApyDisplay,
-  getTerminalDeliveryTxHash,
-  hasFailedSettlement,
-  hasInFlightEarnActions,
-  hashesMatch,
-  inTransitOnlyPositions,
-  isAwaitingFinalize,
-  isCooldownMature,
-  isDeliberateCancel,
-  isEarnRowInFlight,
-  isEarnRowTerminal,
-  isFinalizeInFlight,
-  isLocalEarnTransactionRow,
-  isRecoverPath,
-  isRemoteFailed,
-  isRemoteFailedCancel,
-  isUserCancel,
-  needsManualClaim,
-  needsRecover,
-  pickEarnRowAmount,
-  pickSettleBannerKey,
-  remoteFailedSettlement,
-  remoteFailedStepStatus,
-  resolveSettleStepStatus,
-  resolveStepExplorerChainId,
-  shouldShowRemoteFailedCtas,
-  sumInTransitSharesByShare,
-  unstakeSettlement,
-} from '../../../../../app/[locale]/hemi-earn/_utils'
-import {
-  type EarnPool,
-  type EarnPosition,
-  type EarnSettlement,
-  type EarnTransaction,
-  type EarnTransactionStatusType,
-  type LocalEarnOperation,
-} from '../../../../../app/[locale]/hemi-earn/types'
-
-const baseTx: EarnTransaction = {
-  amountIn: '1000000000000000000',
-  amountOut: null,
-  asset: zeroAddress,
-  automatic: true,
-  cancellationRequested: false,
-  claimTxHash: null,
-  failed: false,
-  kind: 'DEPOSIT',
-  receiver: zeroAddress,
-  recoverTxHash: null,
-  requestedAt: '0',
-  requestId: '0',
-  requestTxHash: `0x${'1'.repeat(64)}`,
-  status: 'PENDING',
-}
-
-const claimHash = `0x${'a'.repeat(64)}` as const
-const recoverHash = `0x${'b'.repeat(64)}` as const
-
-describe('utils', function () {
-  describe('formatApyDisplay', function () {
-    it('should return "< 0.01%" for tiny positive values', function () {
-      expect(formatApyDisplay(0.005)).toBe('< 0.01%')
-    })
-
-    it('should return "< -0.01%" for tiny negative values', function () {
-      expect(formatApyDisplay(-0.005)).toBe('< -0.01%')
-    })
-
-    it('should return "0.00%" for zero', function () {
-      expect(formatApyDisplay(0)).toBe('0.00%')
-    })
-
-    it('should format percentage for negative values beyond the threshold', function () {
-      expect(formatApyDisplay(-5.25)).toBe('-5.25%')
-    })
-
-    it('should format percentage for value equal to 0.01', function () {
-      expect(formatApyDisplay(0.01)).toBe('0.01%')
-    })
-
-    it('should format percentage for larger values', function () {
-      expect(formatApyDisplay(5.25)).toBe('5.25%')
-    })
-  })
-
-  describe('getTerminalDeliveryTxHash', function () {
-    it('returns claimTxHash for FINALIZED', function () {
-      expect(
-        getTerminalDeliveryTxHash({
-          ...baseTx,
-          claimTxHash: claimHash,
-          status: 'FINALIZED',
-        }),
-      ).toBe(claimHash)
-    })
-
-    it('returns recoverTxHash for RECOVERED', function () {
-      expect(
-        getTerminalDeliveryTxHash({
-          ...baseTx,
-          recoverTxHash: recoverHash,
-          status: 'RECOVERED',
-        }),
-      ).toBe(recoverHash)
-    })
-
-    it.each<EarnTransactionStatusType>([
-      'PENDING',
-      'FULFILLED',
-      'CANCELLED',
-      'TX_PENDING',
-      'FAILED',
-    ])('returns undefined for non-terminal status %s', function (status) {
-      expect(
-        getTerminalDeliveryTxHash({
-          ...baseTx,
-          claimTxHash: claimHash,
-          recoverTxHash: recoverHash,
-          status,
-        }),
-      ).toBeUndefined()
-    })
-
-    it('returns undefined when FINALIZED but claimTxHash is null', function () {
-      expect(
-        getTerminalDeliveryTxHash({
-          ...baseTx,
-          claimTxHash: null,
-          status: 'FINALIZED',
-        }),
-      ).toBeUndefined()
-    })
-
-    it('returns undefined when RECOVERED but recoverTxHash is null', function () {
-      expect(
-        getTerminalDeliveryTxHash({
-          ...baseTx,
-          recoverTxHash: null,
-          status: 'RECOVERED',
-        }),
-      ).toBeUndefined()
-    })
-
-    it('returns undefined when tx is undefined', function () {
-      expect(getTerminalDeliveryTxHash(undefined)).toBeUndefined()
-    })
-  })
-
-  describe('findPoolByAsset / findPoolByShare', function () {
-    const shareA = '0x000000000000000000000000000000000000aaaa' as Address
-    const shareB = '0x000000000000000000000000000000000000bbbb' as Address
-    const assetA1 = '0x0000000000000000000000000000000000001111' as Address
-    const assetA2 = '0x0000000000000000000000000000000000002222' as Address
-    const assetB1 = '0x0000000000000000000000000000000000003333' as Address
-    const unknown = '0x0000000000000000000000000000000000009999' as Address
-
-    const makePool = (shareAddress: Address, assets: Address[]) =>
-      ({
-        assets: assets.map(address => ({ address })),
-        shareAddress,
-      }) as unknown as EarnPool
-
-    const pools: EarnPool[] = [
-      makePool(shareA, [assetA1, assetA2]),
-      makePool(shareB, [assetB1]),
-    ]
-
-    it('findPoolByAsset finds the pool whose `assets` includes the address', function () {
-      expect(findPoolByAsset(pools, assetA2)?.shareAddress).toBe(shareA)
-      expect(findPoolByAsset(pools, assetB1)?.shareAddress).toBe(shareB)
-    })
-
-    it('findPoolByAsset returns undefined for an unknown asset', function () {
-      expect(findPoolByAsset(pools, unknown)).toBeUndefined()
-    })
-
-    it('findPoolByShare finds the pool by share address', function () {
-      expect(findPoolByShare(pools, shareB)?.shareAddress).toBe(shareB)
-    })
-
-    it('findPoolByShare returns undefined for an unknown share', function () {
-      expect(findPoolByShare(pools, unknown)).toBeUndefined()
-    })
-  })
-
-  describe('isLocalEarnTransactionRow', function () {
-    it('returns true for a row whose `requestId` is locally-prefixed', function () {
-      expect(
-        isLocalEarnTransactionRow({ ...baseTx, requestId: 'local-1700000000' }),
-      ).toBe(true)
-    })
-
-    it('returns false for a row whose `requestId` is a subgraph numeric id', function () {
-      expect(isLocalEarnTransactionRow({ ...baseTx, requestId: '42' })).toBe(
-        false,
-      )
-    })
-  })
-
-  describe('needsManualClaim', function () {
-    it('is true for a FULFILLED deposit with auto-claim off', function () {
-      expect(
-        needsManualClaim({
-          ...baseTx,
-          automatic: false,
-          kind: 'DEPOSIT',
-          status: 'FULFILLED',
-        }),
-      ).toBe(true)
-    })
-
-    it('is true even when auto-claim is on (auto-finalize reverted leaves it FULFILLED)', function () {
-      expect(
-        needsManualClaim({ ...baseTx, automatic: true, status: 'FULFILLED' }),
-      ).toBe(true)
-    })
-
-    it.each<EarnTransactionStatusType>(['PENDING', 'CANCELLED', 'FINALIZED'])(
-      'is false for non-FULFILLED status %s',
-      function (status) {
-        expect(needsManualClaim({ ...baseTx, automatic: false, status })).toBe(
-          false,
-        )
-      },
-    )
-
-    it('is true for a FULFILLED redeem with auto-claim off (kind-agnostic)', function () {
-      expect(
-        needsManualClaim({
-          ...baseTx,
-          automatic: false,
-          kind: 'REDEEM',
-          status: 'FULFILLED',
-        }),
-      ).toBe(true)
-    })
-  })
-
-  describe('needsRecover', function () {
-    it('is true for a CANCELLED deposit with auto-recover off', function () {
-      expect(
-        needsRecover({
-          ...baseTx,
-          automatic: false,
-          kind: 'DEPOSIT',
-          status: 'CANCELLED',
-        }),
-      ).toBe(true)
-    })
-
-    it('is true even when auto-recover is on (auto-finalize reverted leaves it CANCELLED)', function () {
-      expect(
-        needsRecover({ ...baseTx, automatic: true, status: 'CANCELLED' }),
-      ).toBe(true)
-    })
-
-    it('is false for RECOVERED (already recovered, not actionable)', function () {
-      expect(
-        needsRecover({ ...baseTx, automatic: false, status: 'RECOVERED' }),
-      ).toBe(false)
-    })
-
-    it('is true for a CANCELLED redeem with auto-recover off (kind-agnostic)', function () {
-      expect(
-        needsRecover({
-          ...baseTx,
-          automatic: false,
-          kind: 'REDEEM',
-          status: 'CANCELLED',
-        }),
-      ).toBe(true)
-    })
-  })
-
-  describe('isRecoverPath', function () {
-    it.each<EarnTransactionStatusType>(['CANCELLED', 'RECOVERED'])(
-      'is true for a deposit in status %s regardless of automatic',
-      function (status) {
-        expect(isRecoverPath({ ...baseTx, automatic: true, status })).toBe(true)
-        expect(isRecoverPath({ ...baseTx, automatic: false, status })).toBe(
-          true,
-        )
-      },
-    )
-
-    it.each<EarnTransactionStatusType>(['PENDING', 'FULFILLED', 'FINALIZED'])(
-      'is false for the happy-path status %s',
-      function (status) {
-        expect(isRecoverPath({ ...baseTx, status })).toBe(false)
-      },
-    )
-
-    it.each<EarnTransactionStatusType>(['CANCELLED', 'RECOVERED'])(
-      'is true for a redeem in status %s (kind-agnostic)',
-      function (status) {
-        expect(isRecoverPath({ ...baseTx, kind: 'REDEEM', status })).toBe(true)
-      },
-    )
-  })
-
-  describe('canRetryRow', function () {
-    it('is true for a local FAILED row', function () {
-      expect(
-        canRetryRow({
-          ...baseTx,
-          requestId: 'local-1700000000',
-          status: 'FAILED',
-        }),
-      ).toBe(true)
-    })
-
-    it('is false for a subgraph FAILED row (numeric requestId)', function () {
-      expect(
-        canRetryRow({ ...baseTx, requestId: '42', status: 'FAILED' }),
-      ).toBe(false)
-    })
-
-    it('is false for a non-FAILED local row', function () {
-      expect(
-        canRetryRow({
-          ...baseTx,
-          requestId: 'local-1700000000',
-          status: 'PENDING',
-        }),
-      ).toBe(false)
-    })
-  })
-
-  describe('isRemoteFailed', function () {
-    const remoteFailed: EarnTransaction = {
-      ...baseTx,
-      failed: true,
-      kind: 'REDEEM',
-      requestId: '42',
-      status: 'FAILED',
-    }
-
-    it('is true for a subgraph FAILED redeem flagged failed', function () {
-      expect(isRemoteFailed(remoteFailed)).toBe(true)
-    })
-
-    it('is false for a local FAILED row (retryable from home)', function () {
-      expect(
-        isRemoteFailed({ ...remoteFailed, requestId: 'local-1700000000' }),
-      ).toBe(false)
-    })
-
-    it('is false when the failed flag is not set', function () {
-      expect(isRemoteFailed({ ...remoteFailed, failed: false })).toBe(false)
-    })
-
-    it('is true for a subgraph FAILED deposit flagged failed', function () {
-      expect(isRemoteFailed({ ...remoteFailed, kind: 'DEPOSIT' })).toBe(true)
-    })
-
-    it('is false for a local FAILED deposit', function () {
-      expect(
-        isRemoteFailed({
-          ...remoteFailed,
-          kind: 'DEPOSIT',
-          requestId: 'local-1700000000',
-        }),
-      ).toBe(false)
-    })
-
-    it('is false for a non-FAILED status', function () {
-      expect(isRemoteFailed({ ...remoteFailed, status: 'PENDING' })).toBe(false)
-    })
-
-    it('is false for undefined', function () {
-      expect(isRemoteFailed(undefined)).toBe(false)
-    })
-  })
-
-  describe('isRemoteFailedCancel', function () {
-    const remoteFailed: EarnTransaction = {
-      ...baseTx,
-      failed: true,
-      kind: 'DEPOSIT',
-      requestId: '42',
-      status: 'FAILED',
-    }
-
-    it('is true for a remote-failed row with a CANCEL_REQUEST marker', function () {
-      expect(
-        isRemoteFailedCancel({
-          ...remoteFailed,
-          settlement: { failed: false, kind: 'CANCEL_REQUEST' },
-        }),
-      ).toBe(true)
-    })
-
-    it('is false for a remote-failed row with a RETRY marker', function () {
-      expect(
-        isRemoteFailedCancel({
-          ...remoteFailed,
-          settlement: { failed: false, kind: 'RETRY' },
-        }),
-      ).toBe(false)
-    })
-
-    it('is false for a remote-failed row without a settlement', function () {
-      expect(isRemoteFailedCancel(remoteFailed)).toBe(false)
-    })
-
-    it('is false when the row is not remote-failed', function () {
-      expect(
-        isRemoteFailedCancel({
-          ...remoteFailed,
-          settlement: { failed: false, kind: 'CANCEL_REQUEST' },
-          status: 'PENDING',
-        }),
-      ).toBe(false)
-    })
-  })
-
-  describe('claimRecoverSettlement', function () {
-    it('returns a CLAIM settlement unchanged', function () {
-      const settlement: EarnSettlement = {
-        failed: false,
-        kind: 'CLAIM',
-        txHash: claimHash,
-      }
-      expect(claimRecoverSettlement(settlement)).toBe(settlement)
-    })
-
-    it('returns a RECOVER settlement unchanged', function () {
-      const settlement: EarnSettlement = { failed: true, kind: 'RECOVER' }
-      expect(claimRecoverSettlement(settlement)).toBe(settlement)
-    })
-
-    it('strips a pending CANCEL marker (the cancel signal, not a settlement)', function () {
-      expect(
-        claimRecoverSettlement({ failed: false, kind: 'CANCEL' }),
-      ).toBeUndefined()
-    })
-
-    it('strips a reverted CANCEL marker too', function () {
-      expect(
-        claimRecoverSettlement({ failed: true, kind: 'CANCEL' }),
-      ).toBeUndefined()
-    })
-
-    it('strips an UNSTAKE marker (the Ethereum finalize, not a settlement)', function () {
-      expect(
-        claimRecoverSettlement({ failed: false, kind: 'UNSTAKE' }),
-      ).toBeUndefined()
-      expect(
-        claimRecoverSettlement({ failed: true, kind: 'UNSTAKE' }),
-      ).toBeUndefined()
-    })
-
-    it('strips RETRY and CANCEL_REQUEST markers (Agent-side, not a settlement)', function () {
-      expect(
-        claimRecoverSettlement({ failed: false, kind: 'RETRY' }),
-      ).toBeUndefined()
-      expect(
-        claimRecoverSettlement({ failed: true, kind: 'CANCEL_REQUEST' }),
-      ).toBeUndefined()
-    })
-
-    it('returns undefined for no settlement', function () {
-      expect(claimRecoverSettlement(undefined)).toBeUndefined()
-    })
-  })
-
-  describe('hasFailedSettlement', function () {
-    it('is true when the settlement is flagged failed', function () {
-      expect(
-        hasFailedSettlement({
-          ...baseTx,
-          settlement: { failed: true, kind: 'CLAIM' },
-        }),
-      ).toBe(true)
-    })
-
-    it('is false for a pending (not-yet-failed) settlement', function () {
-      expect(
-        hasFailedSettlement({
-          ...baseTx,
-          settlement: { failed: false, kind: 'RECOVER', txHash: recoverHash },
-        }),
-      ).toBe(false)
-    })
-
-    it('is false when there is no settlement', function () {
-      expect(hasFailedSettlement(baseTx)).toBe(false)
-    })
-
-    it('is false for a reverted CANCEL (the modal owns that retry)', function () {
-      expect(
-        hasFailedSettlement({
-          ...baseTx,
-          settlement: { failed: true, kind: 'CANCEL' },
-        }),
-      ).toBe(false)
-    })
-  })
-
-  describe('findLocalSettlement', function () {
-    const reqHash = `0x${'b'.repeat(64)}` as Hash
-    const settlement: EarnSettlement = {
-      failed: false,
-      kind: 'CLAIM',
-      txHash: claimHash,
-    }
-    const makeLocalOp = (
-      initiateTxHash: string | undefined,
-      withSettlement?: EarnSettlement,
-    ) =>
-      ({
-        initiateTxHash,
-        settlement: withSettlement,
-      }) as unknown as LocalEarnOperation
-
-    it('returns the settlement of the op matching the request tx', function () {
-      const ops = [
-        makeLocalOp(`0x${'9'.repeat(64)}`),
-        makeLocalOp(reqHash, settlement),
-      ]
-      expect(findLocalSettlement(ops, reqHash)).toBe(settlement)
-    })
-
-    it('matches the request tx case-insensitively', function () {
-      const ops = [makeLocalOp(`0x${'B'.repeat(64)}`, settlement)]
-      expect(findLocalSettlement(ops, reqHash)).toBe(settlement)
-    })
-
-    it('returns undefined when no op matches', function () {
-      expect(
-        findLocalSettlement(
-          [makeLocalOp(`0x${'9'.repeat(64)}`, settlement)],
-          reqHash,
-        ),
-      ).toBeUndefined()
-    })
-
-    it('returns undefined when requestTxHash is undefined', function () {
-      expect(
-        findLocalSettlement([makeLocalOp(reqHash, settlement)], undefined),
-      ).toBeUndefined()
-    })
-
-    it('skips ops without an initiateTxHash', function () {
-      expect(
-        findLocalSettlement([makeLocalOp(undefined, settlement)], reqHash),
-      ).toBeUndefined()
-    })
-
-    it('returns undefined when the matching op has no settlement', function () {
-      expect(
-        findLocalSettlement([makeLocalOp(reqHash)], reqHash),
-      ).toBeUndefined()
-    })
-  })
-
-  describe('enrichWithSettlement', function () {
-    const settlement: EarnSettlement = { failed: true, kind: 'RECOVER' }
-
-    it('folds the settlement onto the row', function () {
-      expect(enrichWithSettlement(baseTx, settlement)).toEqual({
-        ...baseTx,
-        settlement,
-      })
-    })
-
-    it('returns the row unchanged when there is no settlement', function () {
-      expect(enrichWithSettlement(baseTx, undefined)).toBe(baseTx)
-    })
-
-    it('returns undefined when the row is undefined', function () {
-      expect(enrichWithSettlement(undefined, settlement)).toBeUndefined()
-    })
-  })
-
-  describe('isEarnRowInFlight', function () {
-    it.each<EarnTransactionStatusType>(['PENDING', 'FULFILLED', 'TX_PENDING'])(
-      'is true for the non-terminal status %s',
-      function (status) {
-        expect(isEarnRowInFlight({ ...baseTx, status })).toBe(true)
-      },
-    )
-
-    it.each<EarnTransactionStatusType>(['FINALIZED', 'RECOVERED'])(
-      'is false for the terminal status %s',
-      function (status) {
-        expect(isEarnRowInFlight({ ...baseTx, status })).toBe(false)
-      },
-    )
-
-    it('is true for a subgraph FAILED row (Agent failed cross-chain; walks to RECOVERED)', function () {
-      expect(
-        isEarnRowInFlight({ ...baseTx, requestId: '40', status: 'FAILED' }),
-      ).toBe(true)
-    })
-
-    it('is false for a local FAILED row (Hemi request tx reverted; terminal, retry from home)', function () {
-      expect(
-        isEarnRowInFlight({
-          ...baseTx,
-          requestId: 'local-1700000000',
-          status: 'FAILED',
-        }),
-      ).toBe(false)
-    })
-
-    it.each([true, false])(
-      'is true for a CANCELLED deposit (automatic=%s — both walk to RECOVERED)',
-      function (automatic) {
-        expect(
-          isEarnRowInFlight({ ...baseTx, automatic, status: 'CANCELLED' }),
-        ).toBe(true)
-      },
-    )
-
-    it.each([true, false])(
-      'is true for a CANCELLED redeem (automatic=%s — now walks to RECOVERED)',
-      function (automatic) {
-        expect(
-          isEarnRowInFlight({
-            ...baseTx,
-            automatic,
-            kind: 'REDEEM',
-            status: 'CANCELLED',
-          }),
-        ).toBe(true)
-      },
-    )
-  })
-
-  describe('hasInFlightEarnActions', function () {
-    const localOp = (initiateTxHash: string | undefined, settled: boolean) =>
-      ({ initiateTxHash, settled }) as unknown as LocalEarnOperation
-
-    it('is false when there are no local ops and no transactions', function () {
-      expect(
-        hasInFlightEarnActions({ localOperations: [], transactions: [] }),
-      ).toBe(false)
-    })
-
-    it('is true when a local op is initiated but not settled yet', function () {
-      expect(
-        hasInFlightEarnActions({
-          localOperations: [localOp(`0x${'9'.repeat(64)}`, false)],
-          transactions: [],
-        }),
-      ).toBe(true)
-    })
-
-    it('is false when the only local op is already settled', function () {
-      expect(
-        hasInFlightEarnActions({
-          localOperations: [localOp(`0x${'9'.repeat(64)}`, true)],
-          transactions: [],
-        }),
-      ).toBe(false)
-    })
-
-    it('is false when a local op has no initiate tx hash', function () {
-      expect(
-        hasInFlightEarnActions({
-          localOperations: [localOp(undefined, false)],
-          transactions: [],
-        }),
-      ).toBe(false)
-    })
-
-    it('is true when a subgraph transaction is still in flight', function () {
-      expect(
-        hasInFlightEarnActions({
-          localOperations: [],
-          transactions: [{ ...baseTx, status: 'PENDING' }],
-        }),
-      ).toBe(true)
-    })
-
-    it('is false when every transaction has reached a terminal status', function () {
-      expect(
-        hasInFlightEarnActions({
-          localOperations: [],
-          transactions: [{ ...baseTx, status: 'FINALIZED' }],
-        }),
-      ).toBe(false)
-    })
-  })
-
-  describe('sumInTransitSharesByShare', function () {
-    const redeemAsset = '0x00000000000000000000000000000000000000c1' as Address
-    const share = '0x00000000000000000000000000000000000000d2' as Address
-    const makePool = (shareAddress: Address, assets: Address[]) =>
-      ({
-        assets: assets.map(address => ({ address })),
-        shareAddress,
-      }) as unknown as EarnPool
-    const pools = [makePool(share, [redeemAsset])]
-    const redeem = (
-      overrides: Partial<EarnTransaction> = {},
-    ): EarnTransaction => ({
-      ...baseTx,
-      amountIn: '2000000000000000000',
-      amountOut: null,
-      asset: redeemAsset,
-      kind: 'REDEEM',
-      status: 'PENDING',
-      ...overrides,
-    })
-
-    it('sums the amountIn of an in-flight redeem under its share', function () {
-      expect(sumInTransitSharesByShare([redeem()], pools)).toEqual({
-        [share.toLowerCase()]: BigInt('2000000000000000000'),
-      })
-    })
-
-    it('counts a CANCELLED redeem (shares still out, pre-recover)', function () {
-      expect(
-        sumInTransitSharesByShare([redeem({ status: 'CANCELLED' })], pools),
-      ).toEqual({ [share.toLowerCase()]: BigInt('2000000000000000000') })
-    })
-
-    it('excludes terminal redeems (RECOVERED / FINALIZED)', function () {
-      expect(
-        sumInTransitSharesByShare(
-          [redeem({ status: 'RECOVERED' }), redeem({ status: 'FINALIZED' })],
-          pools,
-        ),
-      ).toEqual({})
-    })
-
-    it('excludes a FULFILLED redeem whose amountOut has landed', function () {
-      expect(
-        sumInTransitSharesByShare(
-          [redeem({ amountOut: '1900000', status: 'FULFILLED' })],
-          pools,
-        ),
-      ).toEqual({})
-    })
-
-    it('excludes a TX_PENDING redeem (request tx not mined, shares still held)', function () {
-      expect(
-        sumInTransitSharesByShare([redeem({ status: 'TX_PENDING' })], pools),
-      ).toEqual({})
-    })
-
-    it('excludes deposits', function () {
-      expect(
-        sumInTransitSharesByShare([redeem({ kind: 'DEPOSIT' })], pools),
-      ).toEqual({})
-    })
-
-    it('sums multiple in-flight redeems for the same share', function () {
-      expect(
-        sumInTransitSharesByShare(
-          [
-            redeem({ amountIn: '2000000000000000000' }),
-            redeem({ amountIn: '1500000000000000000', requestId: '1' }),
-          ],
-          pools,
-        ),
-      ).toEqual({ [share.toLowerCase()]: BigInt('3500000000000000000') })
-    })
-
-    it('ignores a redeem whose asset maps to no pool', function () {
-      expect(
-        sumInTransitSharesByShare(
-          [redeem({ asset: `0x${'e'.repeat(40)}` })],
-          pools,
-        ),
-      ).toEqual({})
-    })
-  })
-
-  describe('inTransitOnlyPositions', function () {
-    const shareA = '0x00000000000000000000000000000000000000e1' as Address
-    const shareB = '0x00000000000000000000000000000000000000e2' as Address
-    const peggedToken = { decimals: 18 } as unknown as EvmToken
-    const shareToken = { symbol: 'S' } as unknown as EvmToken
-    const makePool = (shareAddress: Address) =>
-      ({
-        assets: [],
-        peggedToken,
-        shareAddress,
-        shareToken,
-      }) as unknown as EarnPool
-    const pools = [makePool(shareA), makePool(shareB)]
-
-    it('synthesizes a zero-balance row for an in-transit share with no position', function () {
-      expect(
-        inTransitOnlyPositions(
-          { [shareA.toLowerCase()]: BigInt(5) },
-          [],
-          pools,
-        ),
-      ).toEqual([
-        {
-          peggedToken,
-          shareAddress: shareA,
-          shareToken,
-          yourDeposit: BigInt(0),
-        },
-      ])
-    })
-
-    it('skips a share already covered by a position', function () {
-      const position = {
-        peggedToken,
-        shareAddress: shareA,
-        shareToken,
-        yourDeposit: BigInt(100),
-      } as unknown as EarnPosition
-      expect(
-        inTransitOnlyPositions(
-          { [shareA.toLowerCase()]: BigInt(5) },
-          [position],
-          pools,
-        ),
-      ).toEqual([])
-    })
-
-    it('drops an in-transit share that maps to no pool', function () {
-      expect(
-        inTransitOnlyPositions(
-          { [`0x${'f'.repeat(40)}`]: BigInt(5) },
-          [],
-          pools,
-        ),
-      ).toEqual([])
-    })
-  })
-
-  describe('isEarnRowTerminal', function () {
-    it.each<EarnTransactionStatusType>(['FINALIZED', 'RECOVERED'])(
-      'is true for the terminal status %s',
-      function (status) {
-        expect(isEarnRowTerminal({ ...baseTx, status })).toBe(true)
-      },
-    )
-
-    it.each<EarnTransactionStatusType>([
-      'CANCELLED',
-      'FAILED',
-      'FULFILLED',
-      'PENDING',
-      'TX_PENDING',
-    ])('is false for the non-terminal status %s', function (status) {
-      expect(isEarnRowTerminal({ ...baseTx, status })).toBe(false)
-    })
-
-    it('is false for a local FAILED row even though it is also out of flight', function () {
-      const localFailed = {
-        ...baseTx,
-        requestId: 'local-1700000000',
-        status: 'FAILED',
-      } as EarnTransaction
-      expect(isEarnRowTerminal(localFailed)).toBe(false)
-      expect(isEarnRowInFlight(localFailed)).toBe(false)
-    })
-  })
-
-  describe('pickEarnRowAmount', function () {
-    const assetToken = { decimals: 8, symbol: 'hemiBTC' } as unknown as EvmToken
-    const shareToken = {
-      decimals: 18,
-      symbol: 'svetBTC',
-    } as unknown as EvmToken
-    const tokens = { assetToken, shareToken }
-
-    it('uses amountIn + assetToken for a DEPOSIT even once amountOut (shares) is set', function () {
-      // Regression: a finalized deposit carries `amountOut` = minted shares
-      // (18-dec). Rendering it against the 8-dec asset token showed 1,000,000
-      // instead of the deposited 0.0001.
-      expect(
-        pickEarnRowAmount(
-          {
-            ...baseTx,
-            amountIn: '10000',
-            amountOut: '100000000000000',
-            kind: 'DEPOSIT',
-            status: 'FINALIZED',
-          },
-          tokens,
-        ),
-      ).toEqual({ rawAmount: '10000', token: assetToken })
-    })
-
-    it('uses amountIn + assetToken for a DEPOSIT with no amountOut yet', function () {
-      expect(
-        pickEarnRowAmount(
-          { ...baseTx, amountIn: '10000', amountOut: null, kind: 'DEPOSIT' },
-          tokens,
-        ),
-      ).toEqual({ rawAmount: '10000', token: assetToken })
-    })
-
-    it('uses amountIn + shareToken for a REDEEM before it is fulfilled', function () {
-      expect(
-        pickEarnRowAmount(
-          { ...baseTx, amountIn: '5000', amountOut: null, kind: 'REDEEM' },
-          tokens,
-        ),
-      ).toEqual({ rawAmount: '5000', token: shareToken })
-    })
-
-    it('uses amountOut + assetToken for a REDEEM once fulfilled', function () {
-      expect(
-        pickEarnRowAmount(
-          { ...baseTx, amountIn: '5000', amountOut: '7777', kind: 'REDEEM' },
-          tokens,
-        ),
-      ).toEqual({ rawAmount: '7777', token: assetToken })
-    })
-
-    it('returns an undefined token for an in-flight REDEEM while the share token is still loading', function () {
-      expect(
-        pickEarnRowAmount(
-          { ...baseTx, amountIn: '5000', amountOut: null, kind: 'REDEEM' },
-          { assetToken },
-        ),
-      ).toEqual({ rawAmount: '5000', token: undefined })
-    })
-  })
-
-  describe('hashesMatch', function () {
-    const lower = `0x${'a'.repeat(64)}` as const
-    const upper = `0x${'A'.repeat(64)}` as const
-    const other = `0x${'b'.repeat(64)}` as const
-
-    it('matches identical hashes', function () {
-      expect(hashesMatch(lower, lower)).toBe(true)
-    })
-
-    it('matches case-insensitively', function () {
-      expect(hashesMatch(lower, upper)).toBe(true)
-    })
-
-    it('does not match different hashes', function () {
-      expect(hashesMatch(lower, other)).toBe(false)
-    })
-
-    it('returns false when the first hash is undefined', function () {
-      expect(hashesMatch(undefined, lower)).toBe(false)
-    })
-
-    it('returns false when the second hash is undefined', function () {
-      expect(hashesMatch(lower, undefined)).toBe(false)
-    })
-
-    it('returns false when both are undefined', function () {
-      expect(hashesMatch(undefined, undefined)).toBe(false)
-    })
-  })
-
-  describe('resolveSettleStepStatus', function () {
-    const base = {
-      awaitingAction: false,
-      fallback: ProgressStatus.NOT_READY,
-      isComplete: false,
-      settlementFailed: false,
-      settlementTxHash: undefined,
-    }
-    const someHash = `0x${'c'.repeat(64)}` as const
-
-    it('is COMPLETED when complete, over a failed/mining/awaiting settlement', function () {
-      expect(
-        resolveSettleStepStatus({
-          ...base,
-          awaitingAction: true,
-          isComplete: true,
-          settlementFailed: true,
-          settlementTxHash: someHash,
-        }),
-      ).toBe(ProgressStatus.COMPLETED)
-    })
-
-    it('is FAILED when the settlement reverted and is not complete', function () {
-      expect(
-        resolveSettleStepStatus({
-          ...base,
-          awaitingAction: true,
-          settlementFailed: true,
-          settlementTxHash: someHash,
-        }),
-      ).toBe(ProgressStatus.FAILED)
-    })
-
-    it('is PROGRESS while the settlement is mining', function () {
-      expect(
-        resolveSettleStepStatus({
-          ...base,
-          awaitingAction: true,
-          settlementTxHash: someHash,
-        }),
-      ).toBe(ProgressStatus.PROGRESS)
-    })
-
-    it('is READY for an untouched manual settlement', function () {
-      expect(resolveSettleStepStatus({ ...base, awaitingAction: true })).toBe(
-        ProgressStatus.READY,
-      )
-    })
-
-    it('falls back to the caller-provided in-flight status otherwise', function () {
-      expect(
-        resolveSettleStepStatus({ ...base, fallback: ProgressStatus.PROGRESS }),
-      ).toBe(ProgressStatus.PROGRESS)
-      expect(resolveSettleStepStatus(base)).toBe(ProgressStatus.NOT_READY)
-    })
-  })
-
-  describe('pickSettleBannerKey', function () {
-    it.each<[EarnTransaction['kind'], EarnTransactionStatusType, string]>([
-      ['DEPOSIT', 'FULFILLED', 'claim-shares'],
-      ['REDEEM', 'FULFILLED', 'claim-funds'],
-      ['DEPOSIT', 'CANCELLED', 'recover-funds'],
-      ['REDEEM', 'CANCELLED', 'recover-shares'],
-    ])('%s %s → %s', function (kind, status, expected) {
-      expect(pickSettleBannerKey({ ...baseTx, kind, status })).toBe(expected)
-    })
-
-    it.each<EarnTransactionStatusType>([
-      'PENDING',
-      'TX_PENDING',
-      'FINALIZED',
-      'RECOVERED',
-      'FAILED',
-    ])('returns undefined for the non-actionable status %s', function (status) {
-      expect(pickSettleBannerKey({ ...baseTx, status })).toBeUndefined()
-    })
-
-    it('returns undefined for an undefined row', function () {
-      expect(pickSettleBannerKey(undefined)).toBeUndefined()
-    })
-
-    it('returns undefined while a claim/recover is pending (settlement marker)', function () {
-      expect(
-        pickSettleBannerKey({
-          ...baseTx,
-          settlement: { failed: false, kind: 'CLAIM' },
-          status: 'FULFILLED',
-        }),
-      ).toBeUndefined()
-    })
-
-    it('returns undefined after a reverted settlement (try-again state)', function () {
-      expect(
-        pickSettleBannerKey({
-          ...baseTx,
-          settlement: { failed: true, kind: 'RECOVER' },
-          status: 'CANCELLED',
-        }),
-      ).toBeUndefined()
-    })
-
-    it('returns "cancelled" for a user-cancelled redeem (indexed flag)', function () {
-      expect(
-        pickSettleBannerKey({
-          ...baseTx,
-          cancellationRequested: true,
-          kind: 'REDEEM',
-          status: 'CANCELLED',
-        }),
-      ).toBe('cancelled')
-    })
-
-    it('returns undefined once a user-cancelled redeem is RECOVERED', function () {
-      expect(
-        pickSettleBannerKey({
-          ...baseTx,
-          cancellationRequested: true,
-          kind: 'REDEEM',
-          status: 'RECOVERED',
-        }),
-      ).toBeUndefined()
-    })
-
-    it('returns undefined when a cancelled redeem FINALIZED (cancel lost the race)', function () {
-      expect(
-        pickSettleBannerKey({
-          ...baseTx,
-          cancellationRequested: true,
-          kind: 'REDEEM',
-          status: 'FINALIZED',
-        }),
-      ).toBeUndefined()
-    })
-
-    it('returns "cancelled" while a CANCEL marker is pending (before indexing)', function () {
-      expect(
-        pickSettleBannerKey({
-          ...baseTx,
-          kind: 'REDEEM',
-          settlement: { failed: false, kind: 'CANCEL' },
-          status: 'PENDING',
-        }),
-      ).toBe('cancelled')
-    })
-
-    it('shows the recover banner for an Agent failure despite a CANCEL marker', function () {
-      expect(
-        pickSettleBannerKey({
-          ...baseTx,
-          failed: true,
-          kind: 'REDEEM',
-          settlement: { failed: false, kind: 'CANCEL' },
-          status: 'CANCELLED',
-        }),
-      ).toBe('recover-shares')
-    })
-  })
-
-  describe('isUserCancel', function () {
-    it('is true when cancellationRequested and not failed', function () {
-      expect(
-        isUserCancel({
-          ...baseTx,
-          cancellationRequested: true,
-          status: 'CANCELLED',
-        }),
-      ).toBe(true)
-    })
-
-    it('is false for an Agent failure (cancellationRequested but failed)', function () {
-      expect(
-        isUserCancel({
-          ...baseTx,
-          cancellationRequested: true,
-          failed: true,
-          status: 'CANCELLED',
-        }),
-      ).toBe(false)
-    })
-
-    it('is false for an Agent failure even with a lingering CANCEL marker', function () {
-      expect(
-        isUserCancel({
-          ...baseTx,
-          failed: true,
-          settlement: { failed: false, kind: 'CANCEL' },
-          status: 'CANCELLED',
-        }),
-      ).toBe(false)
-    })
-
-    it('is true while a CANCEL marker is pending (bridges the indexing lag)', function () {
-      expect(
-        isUserCancel({
-          ...baseTx,
-          settlement: { failed: false, kind: 'CANCEL' },
-        }),
-      ).toBe(true)
-    })
-
-    it('is false for a reverted CANCEL marker (the modal owns that retry)', function () {
-      expect(
-        isUserCancel({
-          ...baseTx,
-          settlement: { failed: true, kind: 'CANCEL' },
-        }),
-      ).toBe(false)
-    })
-
-    it('is false once RECOVERED, even with cancellationRequested (terminal)', function () {
-      expect(
-        isUserCancel({
-          ...baseTx,
-          cancellationRequested: true,
-          status: 'RECOVERED',
-        }),
-      ).toBe(false)
-    })
-
-    it('is false once FINALIZED with cancellationRequested (cancel lost the race)', function () {
-      expect(
-        isUserCancel({
-          ...baseTx,
-          cancellationRequested: true,
-          kind: 'REDEEM',
-          status: 'FINALIZED',
-        }),
-      ).toBe(false)
-    })
-
-    it('is false for a plain row (no flag, no marker)', function () {
-      expect(isUserCancel(baseTx)).toBe(false)
-    })
-  })
-
-  describe('isDeliberateCancel', function () {
-    it('is true for a user cancel (cancellationRequested, not failed)', function () {
-      expect(
-        isDeliberateCancel({ ...baseTx, cancellationRequested: true }),
-      ).toBe(true)
-    })
-
-    it('is true at the terminal RECOVERED state', function () {
-      expect(
-        isDeliberateCancel({
-          ...baseTx,
-          cancellationRequested: true,
-          kind: 'REDEEM',
-          status: 'RECOVERED',
-        }),
-      ).toBe(true)
-    })
-
-    it('is false for a keeper/Agent-failure recovery (failed)', function () {
-      expect(
-        isDeliberateCancel({
-          ...baseTx,
-          cancellationRequested: true,
-          failed: true,
-          status: 'RECOVERED',
-        }),
-      ).toBe(false)
-    })
-
-    it('is false without a cancel request', function () {
-      expect(isDeliberateCancel(baseTx)).toBe(false)
-    })
-  })
-
-  describe('unstakeSettlement', function () {
-    it('returns an UNSTAKE marker unchanged', function () {
-      const settlement: EarnSettlement = { failed: false, kind: 'UNSTAKE' }
-      expect(unstakeSettlement(settlement)).toBe(settlement)
-    })
-
-    it.each<EarnSettlement['kind']>(['CLAIM', 'RECOVER', 'CANCEL'])(
-      'returns undefined for a %s marker',
-      function (kind) {
-        expect(unstakeSettlement({ failed: false, kind })).toBeUndefined()
-      },
-    )
-
-    it('returns undefined for no settlement', function () {
-      expect(unstakeSettlement(undefined)).toBeUndefined()
-    })
-  })
-
-  describe('isCooldownMature', function () {
-    const matureRedeem: EarnTransaction = {
-      ...baseTx,
-      claimableAt: '1700000000',
-      kind: 'REDEEM',
-      status: 'PENDING',
-    }
-
-    it('is true for a PENDING cooldown redeem once remaining hits 0', function () {
-      expect(isCooldownMature(matureRedeem, 0)).toBe(true)
-    })
-
-    it('is false while the cooldown is still counting down', function () {
-      expect(isCooldownMature(matureRedeem, 120)).toBe(false)
-    })
-
-    it('is false when remaining is undefined (no claimableAt yet)', function () {
-      expect(
-        isCooldownMature({ ...matureRedeem, claimableAt: null }, undefined),
-      ).toBe(false)
-    })
-
-    it('is false once the Agent processed it (return flight)', function () {
-      expect(
-        isCooldownMature({ ...matureRedeem, processedAt: '1700000100' }, 0),
-      ).toBe(false)
-    })
-
-    it.each<EarnTransactionStatusType>(['FULFILLED', 'CANCELLED', 'FINALIZED'])(
-      'is false for the non-PENDING status %s',
-      function (status) {
-        expect(isCooldownMature({ ...matureRedeem, status }, 0)).toBe(false)
-      },
-    )
-
-    it('is false for a deposit', function () {
-      expect(isCooldownMature({ ...matureRedeem, kind: 'DEPOSIT' }, 0)).toBe(
-        false,
-      )
-    })
-
-    it('is false while the user is cancelling (cancellationRequested)', function () {
-      expect(
-        isCooldownMature({ ...matureRedeem, cancellationRequested: true }, 0),
-      ).toBe(false)
-    })
-
-    it('is false while a CANCEL marker is pending', function () {
-      expect(
-        isCooldownMature(
-          { ...matureRedeem, settlement: { failed: false, kind: 'CANCEL' } },
-          0,
-        ),
-      ).toBe(false)
-    })
-  })
-
-  describe('isAwaitingFinalize', function () {
-    const awaiting: EarnTransaction = {
-      ...baseTx,
-      claimableAt: '1700000000',
-      kind: 'REDEEM',
-      status: 'PENDING',
-    }
-
-    it('is true for a PENDING cooldown redeem not yet finalized', function () {
-      expect(isAwaitingFinalize(awaiting)).toBe(true)
-    })
-
-    it('is false once the Agent processed it (bridging back)', function () {
-      expect(
-        isAwaitingFinalize({ ...awaiting, processedAt: '1700000100' }),
-      ).toBe(false)
-    })
-
-    it('is false while the user is cancelling', function () {
-      expect(
-        isAwaitingFinalize({ ...awaiting, cancellationRequested: true }),
-      ).toBe(false)
-    })
-
-    it('is false for an instant redeem (no claimableAt)', function () {
-      expect(isAwaitingFinalize({ ...awaiting, claimableAt: null })).toBe(false)
-    })
-
-    it('is false for a deposit', function () {
-      expect(isAwaitingFinalize({ ...awaiting, kind: 'DEPOSIT' })).toBe(false)
-    })
-  })
-
-  describe('isFinalizeInFlight', function () {
-    it('is true when the Agent processed it (processedAt set)', function () {
-      expect(isFinalizeInFlight({ ...baseTx, processedAt: '1700000100' })).toBe(
-        true,
-      )
-    })
-
-    it('is true while a non-failed UNSTAKE marker is mining', function () {
-      expect(
-        isFinalizeInFlight({
-          ...baseTx,
-          settlement: { failed: false, kind: 'UNSTAKE' },
-        }),
-      ).toBe(true)
-    })
-
-    it('is false for a failed UNSTAKE marker (real revert)', function () {
-      expect(
-        isFinalizeInFlight({
-          ...baseTx,
-          settlement: { failed: true, kind: 'UNSTAKE' },
-        }),
-      ).toBe(false)
-    })
-
-    it('ignores a CLAIM marker', function () {
-      expect(
-        isFinalizeInFlight({
-          ...baseTx,
-          settlement: { failed: false, kind: 'CLAIM' },
-        }),
-      ).toBe(false)
-    })
-
-    it('is false with no processedAt and no marker', function () {
-      expect(isFinalizeInFlight(baseTx)).toBe(false)
-    })
-
-    it('is false for undefined', function () {
-      expect(isFinalizeInFlight(undefined)).toBe(false)
-    })
-  })
-
-  describe('remoteFailedSettlement', function () {
-    it('returns a RETRY marker unchanged', function () {
-      const settlement: EarnSettlement = { failed: false, kind: 'RETRY' }
-      expect(remoteFailedSettlement(settlement)).toBe(settlement)
-    })
-
-    it('returns a CANCEL_REQUEST marker unchanged', function () {
-      const settlement: EarnSettlement = {
-        failed: true,
-        kind: 'CANCEL_REQUEST',
-      }
-      expect(remoteFailedSettlement(settlement)).toBe(settlement)
-    })
-
-    it('ignores claim/recover/unstake markers', function () {
-      expect(
-        remoteFailedSettlement({ failed: false, kind: 'CLAIM' }),
-      ).toBeUndefined()
-      expect(
-        remoteFailedSettlement({ failed: false, kind: 'UNSTAKE' }),
-      ).toBeUndefined()
-    })
-
-    it('returns undefined for no settlement', function () {
-      expect(remoteFailedSettlement(undefined)).toBeUndefined()
-    })
-  })
-
-  describe('remoteFailedStepStatus', function () {
-    it('is FAILED when ready and nothing in flight', function () {
-      expect(remoteFailedStepStatus(true, undefined)).toBe(
-        ProgressStatus.FAILED,
-      )
-    })
-
-    it('is in progress while a retry/cancel is in flight', function () {
-      expect(
-        remoteFailedStepStatus(true, { failed: false, kind: 'RETRY' }),
-      ).toBe(ProgressStatus.PROGRESS)
-      expect(
-        remoteFailedStepStatus(true, { failed: false, kind: 'CANCEL_REQUEST' }),
-      ).toBe(ProgressStatus.PROGRESS)
-    })
-
-    it('is FAILED again once the retry/cancel reverted', function () {
-      expect(
-        remoteFailedStepStatus(true, { failed: true, kind: 'RETRY' }),
-      ).toBe(ProgressStatus.FAILED)
-    })
-
-    it('is in progress during the grace (not ready yet)', function () {
-      expect(remoteFailedStepStatus(false, undefined)).toBe(
-        ProgressStatus.PROGRESS,
-      )
-    })
-  })
-
-  describe('resolveStepExplorerChainId', function () {
-    const fallbackChainId = 999
-    const recoveryHash = '0xaaa' as Hash
-
-    it('is undefined when there is no tx hash', function () {
-      expect(
-        resolveStepExplorerChainId({
-          fallbackChainId,
-          settlement: {
-            failed: false,
-            kind: 'CANCEL_REQUEST',
-            txHash: recoveryHash,
-          },
-          txHash: undefined,
-        }),
-      ).toBeUndefined()
-    })
-
-    it('links a remote-failed retry/cancel tx to mainnet', function () {
-      expect(
-        resolveStepExplorerChainId({
-          fallbackChainId,
-          settlement: {
-            failed: false,
-            kind: 'CANCEL_REQUEST',
-            txHash: recoveryHash,
-          },
-          txHash: recoveryHash,
-        }),
-      ).toBe(mainnet.id)
-      expect(
-        resolveStepExplorerChainId({
-          fallbackChainId,
-          settlement: { failed: false, kind: 'RETRY', txHash: recoveryHash },
-          txHash: recoveryHash,
-        }),
-      ).toBe(mainnet.id)
-    })
-
-    it('uses the fallback chain when the hash is not the recovery tx', function () {
-      expect(
-        resolveStepExplorerChainId({
-          fallbackChainId,
-          settlement: {
-            failed: false,
-            kind: 'CANCEL_REQUEST',
-            txHash: recoveryHash,
-          },
-          txHash: '0xbbb' as Hash,
-        }),
-      ).toBe(fallbackChainId)
-    })
-
-    it('uses the fallback chain for a non-remote-failed settlement', function () {
-      expect(
-        resolveStepExplorerChainId({
-          fallbackChainId,
-          settlement: { failed: false, kind: 'CLAIM', txHash: recoveryHash },
-          txHash: recoveryHash,
-        }),
-      ).toBe(fallbackChainId)
-    })
-
-    it('uses the fallback chain when there is no settlement', function () {
-      expect(
-        resolveStepExplorerChainId({
-          fallbackChainId,
-          settlement: undefined,
-          txHash: recoveryHash,
-        }),
-      ).toBe(fallbackChainId)
-    })
-  })
-
-  describe('shouldShowRemoteFailedCtas', function () {
-    const stuckTx: EarnTransaction = {
-      ...baseTx,
-      failed: true,
-      kind: 'REDEEM',
-      receivedAt: '1000',
-      requestId: '42',
-      retryCount: 0,
-      status: 'FAILED',
-    }
-
-    it('is false when not stuck on-chain', function () {
-      expect(
-        shouldShowRemoteFailedCtas({
-          category: 'slippage',
-          isStuck: false,
-          nowSec: 999999,
-          tx: stuckTx,
-        }),
-      ).toBe(false)
-    })
-
-    it('shows immediately for slippage (needs a user call)', function () {
-      expect(
-        shouldShowRemoteFailedCtas({
-          category: 'slippage',
-          isStuck: true,
-          nowSec: 1001,
-          tx: stuckTx,
-        }),
-      ).toBe(true)
-    })
-
-    it('shows once a keeper retry already failed', function () {
-      expect(
-        shouldShowRemoteFailedCtas({
-          category: 'gas',
-          isStuck: true,
-          nowSec: 1001,
-          tx: { ...stuckTx, retryCount: 1 },
-        }),
-      ).toBe(true)
-    })
-
-    it('shows after the grace elapses', function () {
-      expect(
-        shouldShowRemoteFailedCtas({
-          category: 'gas',
-          isStuck: true,
-          nowSec: 1121,
-          tx: stuckTx,
-        }),
-      ).toBe(true)
-    })
-
-    it('stays hidden within the grace window', function () {
-      expect(
-        shouldShowRemoteFailedCtas({
-          category: 'gas',
-          isStuck: true,
-          nowSec: 1060,
-          tx: stuckTx,
-        }),
-      ).toBe(false)
-    })
-
-    it('stays hidden with no receivedAt and no retries', function () {
-      expect(
-        shouldShowRemoteFailedCtas({
-          category: 'unknown',
-          isStuck: true,
-          nowSec: 999999,
-          tx: { ...stuckTx, receivedAt: null },
-        }),
-      ).toBe(false)
-    })
-  })
-})
